% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_multiattribute.R
\name{as_polytope}
\alias{as_polytope}
\title{Transform Pattern of Predictions to Polytope}
\usage{
as_polytope(strategy)
}
\arguments{
\item{strategy}{a decision strategy returned by \code{\link{predict_multiattribute}}.}
}
\value{
a list containing the matrix \code{A} and the vector \code{b}
     that define a polytope via \code{A*x <= b}.
}
\description{
Transforms ordered item-type predictions to polytope definition.
This allows to use Monte-Carlo methods to compute the Bayes factor
if the number of item types is large (\code{\link{compute_bf}}).
}
\details{
Note: Only works for models without guessing predictions and
      without equality constraints (i.e., requires separate error probabilities per item type)
}
\examples{
# define a strategy:
strat <- list(pattern = c(5,-2,3,4),  # B,A,B,B
              c = .5, ordered = TRUE,
              prior = c(1,1))
as_polytope(strat)
}
