% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_encompassing.R
\name{bf_binom}
\alias{bf_binom}
\alias{bf_multinom}
\title{Bayes Factor for Linear Order Constraints}
\usage{
bf_binom(k, n, A, b, V, map, prior = c(1, 1), ...)

bf_multinom(k, options, A, b, V, prior = rep(1, sum(options)), ...)
}
\arguments{
\item{k}{the number of Option B choices.
The default \code{k=n=rep(0,ncol(A))} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the free parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{V}{a matrix of vertices (one per row) that define the polytope of
admissible parameters as the convex hull over these points
(if provided, \code{A} and \code{b} are ignored).
Similar as for \code{A}, columns of \code{V} omit the last value for each
multinomial condition (e.g., a1,a2,a3,b1,b2 becomes a1,a2,b1).
Note that this method is comparatively slow since it solves linear-programming problems
to test whether a point is inside  a polytope (Fukuda, 2004) or to run the Gibbs sampler.}

\item{map}{optional: numeric vector of the same length as \code{k} with integers
mapping the frequencies \code{k} to the free parameters/columns of \code{A}/\code{V},
thereby allowing for equality constraints (e.g., \code{map=c(1,1,2,2)}).
Reversed probabilities \code{1-p} are coded by negative integers.
Guessing probabilities of .50 are encoded by zeros. The default assumes
different parameters for each item type: \code{map=1:ncol(A)}}

\item{prior}{a vector with two positive numbers defining the shape parameters
of the beta prior distributions for each binomial rate parameter.}

\item{...}{further arguments passed to \code{\link{count_binom}}/\code{\link{count_multinom}} (e.g., \code{M}, \code{steps}).}

\item{options}{the number of choice options per item type,
e.g., \code{c(3,2)} for a ternary and binary item.
 The sum of \code{options} must be equal to the length of \code{k}.}
}
\value{
a matrix with estiamtes (first column) and standard errors (second column) for the Bayes factor and log BF of the order-constrained vs. encompassing (\code{bf_0e}) and the encompassing vs. order-constrained (\code{bf_e0}) model
}
\description{
Computes the Bayes factor for linear order-constraints (specified via: A*x <= b)
for product-binomial/-multinomial models.
}
\details{
For more control, use \code{\link{count_binom}} to specifiy how many samples
should be drawn from the prior and posterior, respectively.
This is especially recommended if the same prior distribution (and thus the same integral)
is used for computing BFs for multiple data sets that differ only in \code{k} and \code{n}.
In this case, the prior integral can be computed once with high precision, whereas the
posterior integral needs to be computed separately for each unique vector \code{k}.
}
\examples{
k <- c(0, 3, 2, 5, 3, 7)
n <- rep(10, 6)

# linear order constraint:
# prob_1 < prob_2 < .... < .50
A <- matrix(c(1, -1, 0, 0, 0, 0,
              0, 1, -1, 0, 0, 0,
              0, 0, 1, -1, 0, 0,
              0, 0, 0, 1, -1, 0,
              0, 0, 0, 0, 1, -1,
              0, 0, 0, 0, 0, 1),
            ncol = 6, byrow = TRUE)
b <- c(0, 0, 0, 0, 0, .50)

# Bayes factor: unconstrained vs. constrained
bf_binom(k, n, A, b, prior=c(1, 1), M=5e4)
bf_binom(k, n, A, b, prior=c(1, 1), M=5000, steps=c(2,4,5))
bf_binom(k, n, A, b, prior=c(1, 1), M=1000, cmin = 500)
}
\references{
Karabatsos, G. (2005). The exchangeable multinomial model as an approach to testing deterministic axioms of choice and measurement. Journal of Mathematical Psychology, 49(1), 51-69. \url{https://doi.org/10.1016/j.jmp.2004.11.001}

Regenwetter, M., Davis-Stober, C. P., Lim, S. H., Guo, Y., Popova, A., Zwilling, C., … Messner, W. (2014). QTest: Quantitative testing of theories of binary choice. Decision, 1(1), 2-34. \url{https://doi.org/10.1037/dec0000007}
}
\seealso{
\code{\link{count_binom}} and \code{\link{count_multinom}} for
    for more control on the number of prior/posterior samples and
    \code{\link{bf_nonlinear}} for nonlinear order constraints.
}
