% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_encompassing.R
\name{bf_binomial}
\alias{bf_binomial}
\alias{bf_multinomial}
\title{Encompassing Bayes Factor}
\usage{
bf_binomial(k, n, A, b, V, prior = c(1, 1), M = 5e+05, steps,
  batch = 10000)

bf_multinomial(k, options, A, b, V, prior = rep(1, sum(options)), M = 5e+05,
  steps, batch = 10000)
}
\arguments{
\item{k}{the number of Option B choices.
The default \code{k=0} an \code{n=0} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the free parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{V}{a matrix of vertices (one per row) that define the polytope of
admissible parameters as the convex hull over these points
(if provided, \code{A} and \code{b} are ignored). Note that this method
is comparatively slow since it solves a linear-programming problem for each point (Fukuda, 2004).}

\item{prior}{a vector with two positive numbers defining the shape parameters of the beta prior distributions for each binomial rate parameter.}

\item{M}{number of posterior samples drawn from the encompassing model}

\item{steps}{integer vector that indicates at which rows the matrix \code{A} is split for a stepwise computation of the Bayes factor (see details). In this case, \code{M} can be a vector with the number of samples drawn in each step from the (partially) order-constrained models using Gibbs sampling}

\item{batch}{size of the batches into which computations are split to reduce memory load}

\item{options}{the number of choice options per item type, e.g., \code{c(3,2)} for a ternary and binary item.
The sum of \code{options} must be equal to the length of \code{k}.}
}
\value{
a matrix with estiamtes (first column) and standard errors (second column) for the Bayes factor and log BF of the order-constrained vs. encompassing (\code{bf_0e}) and the encompassing vs. order-constrained (\code{bf_e0}) model
}
\description{
Computes the Bayes factor for order-constrained product-binomial/-multinomial models
that are specified as a polytope via:  A*x <= b (see details).
}
\details{
For more control, use \code{\link{count_binomial}} to specifiy how many samples
should be drawn from the prior and posterior, respectively.
This is especially recommended if the same prior distribution (and thus the same integral)
is used for computing BFs for multiple data sets that differ only in \code{k} and \code{n}.
}
\examples{
k <- c(0, 3, 2, 5, 3, 7)
n <- rep(10, 6)

# linear order constraint:
# theta1 < theta2 < .... < .50
A <- matrix(c(1, -1, 0, 0, 0, 0,
              0, 1, -1, 0, 0, 0,
              0, 0, 1, -1, 0, 0,
              0, 0, 0, 1, -1, 0,
              0, 0, 0, 0, 1, -1,
              0, 0, 0, 0, 0, 1),
            ncol = 6, byrow = TRUE)
b <- c(0, 0, 0, 0, 0, .50)

# check whether vectors are in polytope:
A \%*\% c(.05, .1, .12, .16, .19, .23) <= b
A \%*\% c(.05, .3, .12, .16, .19, .53) <= b

# Bayes factor: unconstrained vs. constrained
bf_binomial(k, n, A, b, prior=c(1, 1), M=2e5)
bf_binomial(k, n, A, b, prior=c(1, 1), M=1e4, steps=c(2,4,5))
}
\references{
Karabatsos, G. (2005). The exchangeable multinomial model as an approach to testing deterministic axioms of choice and measurement. Journal of Mathematical Psychology, 49(1), 51-69. \url{https://doi.org/10.1016/j.jmp.2004.11.001}

Regenwetter, M., Davis-Stober, C. P., Lim, S. H., Guo, Y., Popova, A., Zwilling, C., … Messner, W. (2014). QTest: Quantitative testing of theories of binary choice. Decision, 1(1), 2-34. \url{https://doi.org/10.1037/dec0000007}
}
