% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_nonlinear.R
\name{bf_nonlinear}
\alias{bf_nonlinear}
\title{Bayes Factor for Nonlinear Order Constraints}
\usage{
bf_nonlinear(k, options, inside, prior = rep(1, sum(options)),
  M = 10000)
}
\arguments{
\item{k}{the number of choices for each alternative ordered by item type, e.g.
\code{c(a1,a2,a3, b1,b2)}.
The default \code{k=0} is equivalent to sampling from the prior.}

\item{options}{the number of choice options per item type,
e.g., \code{c(3,2)} for a ternary and binary item.
 The sum of \code{options} must be equal to the length of \code{k}.}

\item{inside}{a function that takes a vector with probabilities
\code{p=c(p11,p12,p13,  p21,p22,...)} as input and returns \code{1} or \code{TRUE}
if the order constraints are satisfied and \code{0} or \code{FALSE} if not.}

\item{prior}{the prior parameters of the Dirichlet-shape parameters.
Must have the same length as \code{k}.}

\item{M}{number of posterior samples drawn from the encompassing model}
}
\description{
Draws prior and posterior samples from product-multinomial model and computes the
Bayes factor for a user-specified nonlinear order constraint defined via an
indicator function of the parameters: (p11,p12,p13,  p21,p22,...)
}
\examples{
### 2x2x2 continceny table (Klugkist & Hojtink, 2007)
# (defendant's race) x (victim's race) x (death penalty)
# index: 0 = white/white/yes  ; 1 = black/black/no

# p = (p000,p001,  p010,p011,  p100,p101,  p110,p111)
k <- c(19,132,     0, 9,       11, 52,     6,97)

# Model2: p000*p101 < p100*p001 & p010*p111 < p110*p011
model2 <- function(x)
   x[1]*x[6] < x[5]*x[2] & x[3]*x[8] < x[7]*x[4]
bf_nonlinear(k, 8, model2, M=1e5)  # K&H07: bf_0u=1.62
}
\references{
Klugkist, I., & Hoijtink, H. (2007). The Bayes factor for inequality and about equality constrained models. Computational Statistics & Data Analysis, 51(12), 6367-6379. \url{https://doi.org/10.1016/j.csda.2007.01.024}
}
