% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml.R
\name{compute_cnml}
\alias{compute_cnml}
\title{Compute NML Complexity}
\usage{
compute_cnml(prediction, n, c = 0.5, luck = c(1, 1), n.fit = 3,
  cores = 1)
}
\arguments{
\item{prediction}{a vector of strategy predictions. See \code{\link{predict_multiattribute}}}

\item{n}{vector with the number of choices per item type}

\item{c}{upper threshold of probabilities}

\item{luck}{parameters of luckiness function of LNML (i.e., parameters of a beta distribution). \code{luck = c(1.5, 1.5)} is equivalent to uniform prior for the Bayes factor}

\item{n.fit}{number of repeated fitting runs with random starting values}

\item{cores}{number of processing units to be used}
}
\description{
Enumerates discrete data space and adds the maximum likelihood values.
}
