% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{error_to_prob}
\alias{error_to_prob}
\title{Choice Probabilities Implied by Strategy Predictions}
\usage{
error_to_prob(error, strategy)
}
\arguments{
\item{error}{parameter vector of error probabilities (depending on \code{pattern}).
A single value for deterministic strategies and an
order vector of probabilities for probabilistic strategies (cf. examples).}

\item{strategy}{a list that defines the predictions of a strategy, see\code{\link{predict_multiattribute}}}
}
\description{
Returns a vector with probabilities of choosing Option B for a predicted
pattern and corresponding error probabilities.
}
\examples{
# Predicted:  B,B,[guess],A
# (deterministic: with constant error = .10)
s1 <- list(pattern = c(1, 1, 0, -1),
           ordered = FALSE, c = .1, prior = c(1,1))
error_to_prob(error = .10, s1)

# Predicted:  B,B,A,[guess]
# (probabilistic: with ordered errors e1<e2<e3<.50)
s2 <- list(pattern = c(1, 2, -3, 0),
           ordered = TRUE, c = .5, prior = c(1,1))
error_to_prob(error = c(.10, .15,.20), pattern2, ordered = TRUE)
}
