% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{error_to_probB}
\alias{error_to_probB}
\title{Choice Probabilities Implied by Strategy Predictions}
\usage{
error_to_probB(error, prediction)
}
\arguments{
\item{error}{parameter vector of error probabilities (depending on \code{prediction}).
A single value for deterministic strategies and an
order vector of probabilities for probabilistic strategies (cf. examples).}
}
\description{
Returns a vector with probabilities of choosing Option B for a given set
of predictions and corresponding error probabilities.
}
\examples{
# Predicted:  B,B,[guess],A
# (deterministic: with constant error = .10)
pred <- c(1, 1, 0, -1)
error_to_probB(error = .10, pred)

# Predicted:  B,B,A,[guess]
# (probabilistic: with ordered errors e1<e2<e3)
pred2 <- c(1, 2, -3, 0)
error_to_probB(error = c(.10, .15,.20), pred2)
}
