% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{get_prob_B}
\alias{get_prob_B}
\title{Get Probabilities Predicted by Strategy}
\usage{
get_prob_B(par, prediction)
}
\arguments{
\item{par}{parameter vector (i.e., error probabilities)}

\item{prediction}{a vector of strategy predictions. See \code{\link{get_prediction}}}
}
\description{
Returns a vector with probabilities of choosing Option B for the different item types.
}
