% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilbig2014}
\alias{hilbig2014}
\title{Data: Multiattribute Decisions (Hilbig & Moshagen, 2014)}
\format{A data frame 3 variables:
\describe{
  \item{\code{B1}}{Frequency of choosing Option B for Item Type 1}
  \item{\code{B2}}{Frequency of choosing Option B for Item Type 2}
  \item{\code{B3}}{Frequency of choosing Option B for Item Type 3}
}}
\usage{
hilbig2014
}
\description{
Dataset with multiattribute decisions across 3 item types (Hilbig & Moshagen, 2014).
}
\details{
Each participant made 32 choices for each of 3 item types with four cues (with validities .9, .8, .7, and .6).
}
\examples{
data(hilbig2014)
head(hilbig2014)

\dontrun{
n <- rep(32, 3)
preds <- list(WADDD = c(-1, 1, -1),
              TTB =   c(-1, -1, -1),
              EQW =   c(-1, 1, 0),
              GUESS = c(0, 0, 0))
cnmls <- compute_cnml(preds, n, c = .5)
cnml_baseline <- compute_cnml(1:3, n, c = 1)
select_nml(hilbig2014[1:5,], n,
           c(list(baseline=cnml_baseline), cnmls))
}
}
\references{
Hilbig, B. E., & Moshagen, M. (2014). Generalized outcome-based strategy classification: Comparing deterministic and probabilistic choice models. Psychonomic Bulletin & Review, 21(6), 1431-1443. https://doi.org/10.3758/s13423-014-0643-0
}
\keyword{datasets}
