% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_to_inequalities.R
\name{ineq_probabilities}
\alias{ineq_probabilities}
\title{Inequality Constraints for Product-Multinomial Probabilities}
\usage{
ineq_probabilities(options, A = NULL, b = NULL, nonneg = FALSE)
}
\arguments{
\item{options}{the number of choice options per item type, e.g., \code{c(3,2)} for a ternary and binary item.
The sum of \code{options} must be equal to the length of \code{k}.}

\item{A}{a matrix defining the convex polytope via A*x <= b.
The columns of A do not include the last choice option per item type and
thus the number of columns must be equal to \code{sum(options-1)} (e.g., the column order of \code{A}
for \code{k = c(a1,a2,a2, b1,b2)} is \code{c(a1,a2, b1)}).}

\item{b}{a vector of the same length as the number of rows of \code{A}}

\item{nonneg}{whether to add constraints that probabilities must be nonnegative}
}
\description{
Get or add inequality constraints that multinomial probabilities are
positive and sum to one for all choice options within each item type.
}
\details{
If \code{A} and \code{b} are provided, the constraints are added to these inequality constraints.
}
\examples{
# three binary and two ternary choices:
options <- c(2,2,2, 3,3)
ineq_probabilities(options)
ineq_probabilities(options, nonneg = TRUE)
}
