% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.R
\name{inside_binomial}
\alias{inside_binomial}
\alias{inside_multinomial}
\title{Check Whether Choice Frequencies are in Polytope}
\usage{
inside_binomial(k, n, A, b, V)

inside_multinomial(k, options, A, b, V)
}
\arguments{
\item{k}{choice frequencies.
For \code{inside_binomial}: per item type (e.g.: a1,b1,c1,..)
For \code{inside_multinomial}: for all choice options ordered by item type
(e.g., for ternary choices: a1,a2,a3, b1,b2,b3,..)}

\item{n}{only for \code{inside_binomial}: number of choices per item type.}

\item{options}{only for \code{inside_multinomial}: number of response options per item type.}
}
\description{
Computes relative choice frequencies and checks whether these are in the polytope defined
via (1) \code{A*x <= b} or (2) by the convex hull of a set of vertices \code{V}.
}
\examples{
############ binomial
# x1<x2<x3<.50:
A <- matrix(c(1,-1,0,
              0,1,-1,
              0,0, 1), ncol=3, byrow=TRUE)
b <- c(0, 0, .50)
k <- c( 0, 1, 5)
n <- c(10,10,10)
inside_binomial(k, n, A, b)

############ multinomial
# two ternary choices:
#     (a1,a2,a3,   b1,b2,b3)
k <- c(1,4,10,     5,9,1)
options <- c(3, 3)
# a1<b1, a2<b2:
A <- matrix(c(1,-1,0, 0,
              0, 0,1,-1), ncol=4, byrow=TRUE)
b <- c(0, 0)
inside_multinomial(k, options, A, b)
}
\seealso{
\code{\link{inside}}
}
