% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_karabatsos.R
\docType{data}
\name{karabatsos2004}
\alias{karabatsos2004}
\title{Data: Item Responses Theory (Karabatsos & Sheu, 2004)}
\format{A matrix with 6 rows and columns:
\describe{
  \item{\code{rs}(i=0,...,5): }{Sum score without the item i}
  \item{\code{item}(j=1,...,6): }{Item number}
}}
\usage{
karabatsos2004
}
\description{
Frequency table with number of correct responses \eqn{y[ij]} on item \eqn{j}
obtained by the \eqn{N[ij]} respondents who have rest score \eqn{R[j] = i}.
}
\examples{
data(karabatsos2004)
head(karabatsos2004)

}
\references{
Karabatsos, G., & Sheu, C.-F. (2004). Order-constrained Bayes inference for dichotomous models of unidimensional nonparametric IRT. Applied Psychological Measurement, 28(2), 110-125. \url{https://doi.org/10.1177/0146621603260678}
}
\seealso{
The polytope for the nonparametric item response theory can be obtained using (see \code{\link{nirt_to_Ab}}).
}
\keyword{datasets}
