% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_estimates.R
\name{ml_binomial}
\alias{ml_binomial}
\alias{ml_multinomial}
\title{Maximum-likelihood Estimate}
\usage{
ml_binomial(k, n, A, b, map, strategy, n.fit = 1)

ml_multinomial(k, options, A, b, n.fit = 1)
}
\arguments{
\item{k}{the number of Option B choices.
The default \code{k=n=rep(0,ncol(A))} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the free parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{map}{optional: numeric vector of the same length as \code{k} with integers mapping the frequencies \code{k} to the free parameters/columns of \code{A}/\code{V}, thereby allowing for equality constraints (e.g., \code{map=c(1,1,2,2)}). Reversed probabilities \code{1-p} are coded by negative integers. Guessing probabilities of .50 are encoded by zeros. The default assumes different parameters for each item type: \code{map=1:ncol(A)}}

\item{strategy}{a list that defines the predictions of a strategy, see\code{\link{strategy_multiattribute}}.}

\item{n.fit}{number of calls to \link[stats]{constrOptim}.}

\item{options}{the number of choice options per item type, e.g., \code{c(3,2)} for a ternary and binary item.
The sum of \code{options} must be equal to the length of \code{k}.}
}
\description{
Get ML estimate for product-binomial/multinomial model with linear inequality constraints.
}
\examples{
# linear order: p1 < p2 < p3 < .50
# (cf. WADDprob in ?strategy_multiattribute)
A <- matrix(c(1, -1,  0,
              0,  1, -1,
              0,  0,  1),
            ncol = 3, byrow = TRUE)
b <- c(0, 0, .50)
ml_binomial(c(4,1,23), 40, A, b)



# probabilistic strategy:  A,A,A,B  [e1<e2<e3<e4<.50]
strat <- list(pattern = c(-1, -2, -3, 4),
           c = .5, ordered = TRUE, prior = c(1,1))
ml_binomial(c(7,3,1, 19), 20, strategy = strat)
}
