% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{post_prob}
\alias{post_prob}
\alias{post_prob.default}
\alias{post_prob.matrix}
\alias{post_prob.data.frame}
\title{Get Posterior Model Probabilities}
\usage{
post_prob(marginal, prior)

\method{post_prob}{default}(marginal, prior)

\method{post_prob}{matrix}(marginal, prior)

\method{post_prob}{data.frame}(marginal, prior)
}
\arguments{
\item{marginal}{vector of log-marginal probabilities}

\item{prior}{vector of prior model probabilities (default: uniform over models). Note that vector is normalized internally to sum to one.}
}
\description{
Computes the posterior model probabilities based on the marginal likelihoods
and the prior model probabilities.
}
