% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprob}
\alias{postprob}
\title{Transform Bayes Factors to Posterior Model Probabilities}
\usage{
postprob(..., prior, beta = c(0.5, 0.5), samples = 3000)
}
\arguments{
\item{...}{an arbitrary number of Bayes-factor objects with the elements \code{count} and \code{M} indicating how often posterior samples where within the admissible parameter space relative to the same encompassing model. Such lists are returned by \code{\link{count_binomial}} or \code{\link{count_multinomial}}.}

\item{prior}{prior model probabilities. Uniform by default.}

\item{beta}{prior parameters of beta distributions for estimating the standard error}

\item{samples}{number of samples from beta distributions used to approximate the standard error. The default is Jeffreys' prior.

The encompassing model is the unconstrained baseline model that assumes a separate,
unconstrained probability for each observable frequency.

The standard error of the (stepwise) encompassing Bayes factor is estimated by sampling
ratios from beta distributions, with parameters defined by the posterior/prior counts
(see Hoijtink, 2011; p. 211).}
}
\description{
Computes posterior model probabilities based on objects returned by
\code{\link{bf_binomial}}, \code{\link{bf_multinomial}} and \code{\link{count_to_bf}}.
}
