% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{ppp_binomial}
\alias{ppp_binomial}
\alias{ppp_multinomial}
\title{Posterior Predictive p-Values}
\usage{
ppp_binomial(theta, k, n, M = 2000)

ppp_multinomial(theta, k, options, M = 2000)
}
\arguments{
\item{theta}{posterior samples of the choice probabilities.
See \code{\link{sampling_binomial}} and \code{\link{sampling_multinomial}}}

\item{k}{the number of Option B choices.
The default \code{k=0} an \code{n=0} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{M}{number of subsamples from \code{theta}}

\item{options}{the number of choice options per item type, e.g., \code{c(3,2)} for a ternary and binary item.
The sum of \code{options} must be equal to the length of \code{k}.}
}
\description{
Uses posterior samples to get posterior-predicted frequencies and compare
the Pearson's X^2 statistic for (1) the observed frequencies vs. (2) the posterior-predicted frequencies.
}
\examples{
# uniform samples:  p<.10
theta <- matrix(runif(300*3, 0, .1), 300)
n <- rep(10, 3)
ppp_binomial(theta, c(1,2,0), n)  # ok
ppp_binomial(theta, c(5,4,3), n)  # misfit

# multinomial (ternary choice)
theta <- matrix(runif(300*2, 0, .05), 300)
ppp_multinomial(theta, c(1,0,9), 3)
}
\references{
Myung, J. I., Karabatsos, G., & Iverson, G. J. (2005). A Bayesian approach to testing decision making axioms. \emph{Journal of Mathematical Psychology, 49}, 205-225. \url{https://doi.org/10.1016/j.jmp.2005.02.004}
}
\seealso{
\code{\link{sampling_binomial}} and \code{\link{sampling_multinomial}} to get posterior samples
}
