% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_multiattribute.R
\name{predict_multiattribute}
\alias{predict_multiattribute}
\alias{predict_multiattribute.default}
\alias{predict_multiattribute.matrix}
\alias{predict_multiattribute.data.frame}
\title{Get Predictions of Common Strategies}
\usage{
predict_multiattribute(cueA, cueB, v, strategy)

\method{predict_multiattribute}{default}(cueA, cueB, v, strategy)

\method{predict_multiattribute}{matrix}(cueA, cueB, v, strategy)

\method{predict_multiattribute}{data.frame}(cueA, cueB, v, strategy)
}
\arguments{
\item{cueA}{cue values of Option A (-1/+1 = negative/positive; 0 = missing).
If a matrix is provided, each row defines one item type.}

\item{cueB}{cue values of Option B (see \code{cueA}).}

\item{v}{cue validities: probabilities that cues lead to correct decision.
Must be of the same length as the number of cues.}

\item{strategy}{strategy label, e.g., \code{"TTB"}, \code{"WADD"}, or \code{"WADDprob"}. See details.}
}
\value{
Predidcted choice pattern are encoded as negative = Option A, positive = Option B, and 0 = guessing.
        Different integers are used if error probabilities are assumed to be different.
}
\description{
Returns a vector of predictions for well-known choice strategies (TTB, WADD)
}
\examples{
ca <- c(1, -1, -1, 1)
cb <- c(-1, 1, -1, -1)
v <- c(.9, .8, .7, .6)
predict_multiattribute(ca, cb, v, "TTBprob")
predict_multiattribute(ca, cb, v, "TTB")
predict_multiattribute(ca, cb, v, "WADDprob")
predict_multiattribute(ca, cb, v, "WADD")
predict_multiattribute(ca, cb, v, "EQW")
predict_multiattribute(ca, cb, v, "GUESS")
}
