% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regenwetter2012}
\alias{regenwetter2012}
\title{Data: Ternary Risky Choices (Regenwetter & Davis-Stober, 2012)}
\format{A matrix with 22 columns:
\describe{
  \item{\code{participant}: }{Participant number}
  \item{\code{gamble_set}: }{Gamble set}
  \item{\code{a>b}: }{Number of times a preferred over b}
  \item{\code{b>a}: }{Number of times b preferred over a}
  \item{\code{a=b}: }{Number of times being indifferent between a and b}
}}
\usage{
regenwetter2012
}
\description{
Raw data with risky choices between all 20 paired comparison of 5 elements a,b,c,d, and e.
Participants could either choose "Option 1", "Option 2", or "indifferent".
Each gamble pair (e.g., a vs. b) was repeated 45 times per participant.
There werre three different gamble sets (see Regenwetter & Davis-Stober, 2012).
}
\details{
Each participant made 40 choices for each of 4 item types with four cues
(with validities .9, .8, .7, and .6).
Individual choice freqeuncies are available as \code{\link{heck2017}}
}
\examples{
data(regenwetter2012)
head(regenwetter2012)

\dontrun{
# check predictions
}
}
\references{
Regenwetter, M., & Davis-Stober, C. P. (2012). Behavioral variability of choices versus structural inconsistency of preferences. Psychological Review, 119(2), 408-416. \url{https://doi.org/10.1037/a0027372}
}
\keyword{datasets}
