% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmultinom.R
\name{rpmultinom}
\alias{rpmultinom}
\alias{rpbinom}
\title{Random Sample for Product-Multinomial Distribution}
\usage{
rpmultinom(theta, n, options)

rpbinom(theta, n)
}
\arguments{
\item{theta}{vector with probability parameters. Must sum to one within each option.
Can be a matrix with one parameter vector per row.
See \code{\link{sampling_binomial}} and \code{\link{sampling_multinomial}}.}

\item{n}{integer vector, specifying the total number of objects for each multinomial distribution.}

\item{options}{number of parameters/categories of each multinomial distribution.}
}
\description{
Generate random samples from independent multinomial distributions.
}
\examples{
# 5 binomials
rpbinom(c(.2, .7, .9, .1, .5), rep(20, 5))

# 2 and 3 options:  [a1,a2,  b1,b2,b3]
rpmultinom(c(.5,.5, .3,.6,.1), c(100, 200), c(2, 3))

theta <- rpdirichlet(5, rep(1,5), c(2,3))
rpmultinom(theta, c(20, 20), c(2, 3))
}
