% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_polytope.R
\name{sampling_binom}
\alias{sampling_binom}
\title{Posterior Sampling for Constrained Binomial Models}
\usage{
sampling_binom(k, n, A, b, map = 1:ncol(A), prior = c(1, 1), M = 5000,
  start, burnin = 10, progress = TRUE)
}
\arguments{
\item{k}{the number of Option B choices.
The default \code{k=n=rep(0,ncol(A))} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the free parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{map}{optional: numeric vector of the same length as \code{k} with integers
mapping the frequencies \code{k} to the free parameters/columns of \code{A}/\code{V},
thereby allowing for equality constraints (e.g., \code{map=c(1,1,2,2)}).
Reversed probabilities \code{1-p} are coded by negative integers.
Guessing probabilities of .50 are encoded by zeros. The default assumes
different parameters for each item type: \code{map=1:ncol(A)}}

\item{prior}{a vector with two positive numbers defining the shape parameters
of the beta prior distributions for each binomial rate parameter.}

\item{M}{number of posterior samples}

\item{start}{only relevant if \code{steps} is defined or \code{cmin>0}:
a vector with starting values in the interior of the polytope.
If missing, an approximate maximum-likelihood estimate is used.}

\item{burnin}{number of burnin samples that are discarded. can be chosen to be
small if the maxmimum-likelihood estimate is used as a starting value.}

\item{progress}{whether a progress bar should be shown.}
}
\value{
a matrix with posterior samples for the probabilities to choose Option B for each item type
}
\description{
Uses Gibbs sampling to draw posterior samples for
binomial models with linear inequality-constraints.
}
\details{
Can be used to obtain posterior samples for binary random utility models that
assume a mixture over predefined preference orders/vertices that jointly define
a convex polytope via (A * x <= b).
}
\examples{
A <- matrix(c(1, 0, 0,   # x1 < .50
              1, 1, 1,   # x1+x2+x3 < 1
              0, 2, 2,   # 2*x2+2*x3 < 1
              0, -1, 0,  # x2 > .2
              0, 0, 1),  # x3 < .1
            ncol = 3, byrow = TRUE)
b <- c(.5, 1, 1, -.2, .1)
samp <- sampling_binom(c(5,12,7), c(20,20,20), A, b)
head(samp)
apply(samp, 2, plot, type = "l", ylim = c(0,.6))
}
\references{
Myung, J. I., Karabatsos, G., & Iverson, G. J. (2005). A Bayesian approach to testing decision making axioms. \emph{Journal of Mathematical Psychology, 49}, 205-225. \url{https://doi.org/10.1016/j.jmp.2005.02.004}
}
\seealso{
\code{\link{count_binom}}
}
