% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_polytope.R
\name{sampling_binomial}
\alias{sampling_binomial}
\title{Posterior Samples for Ordered-Constrained Product-Binomial Models}
\usage{
sampling_binomial(k, n, A, b, prior = c(1, 1), M = 5000, start,
  burnin = 10)
}
\arguments{
\item{k}{the number of Option B choices.
The default \code{k=0} an \code{n=0} is equivalent to sampling from the prior.}

\item{n}{the number of choices per item type.}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the free parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{prior}{a vector with two positive numbers defining the shape parameters of the beta prior distributions for each binomial rate parameter.}

\item{M}{number of posterior samples}

\item{start}{starting vectors. Must be in the convex polytope (that is, \code{A*start <= b})}

\item{burnin}{number of burnin samples that are discarded}
}
\value{
a matrix with posterior samples for the probabilities to choose Option B for each item type
}
\description{
Uses Gibbs sampling to draw posterior samples for order-constrained models of binary choice.
}
\details{
Can be used to obtain posterior samples for binary random utility models that assume a mixture over predefined
preference orders/vertices that jointly define a convex polytope via (A * x <= b).
}
\examples{
A <- matrix(c(1, 0, 0,   # x1 < .50
              1, 1, 1,   # x1+x2+x3 < 1
              0, 2, 2,   # 2*x2+2*x3 < 1
              0, -1, 0,  # x2 > .2
              0, 0, 1),  # x3 < .1
            ncol = 3, byrow = TRUE)
b <- c(.5, 1, 1, -.2, .1)
samp <- sampling_binomial(c(5,12,7), c(20,20,20), A, b)
head(samp)
colMeans(samp)
apply(samp, 2, plot, type = "l", ylim = c(0,.6))
}
\references{
Myung, J. I., Karabatsos, G., & Iverson, G. J. (2005). A Bayesian approach to testing decision making axioms. \emph{Journal of Mathematical Psychology, 49}, 205-225. \url{https://doi.org/10.1016/j.jmp.2005.02.004}
}
\seealso{
\code{\link{count_binomial}}
}
