% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_vertex.R
\name{sampling_V}
\alias{sampling_V}
\title{Gibbs Sampler for Polytope based on Vertex representation}
\usage{
sampling_V(k, options, V, prior = rep(1, sum(options)), M = 5000, start,
  burnin = 10, bisection = 8, progress = TRUE)
}
\description{
Gibbs Sampler for Polytope based on Vertex representation
}
