% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_bf.R
\name{se_bf}
\alias{se_bf}
\title{Standard Error for Bayes Factor Estimate}
\usage{
se_bf(npost, nprior, M, prior = c(0.5, 0.5), samples = 1000)
}
\arguments{
\item{npost}{number of posterior samples that satisfy order constraints
(in the stepwise procedure: a vector of integers)}

\item{M}{total number of posterior samples that have been drawn to compute the BF.
In the stepwise procedure: if the length of \code{M} is one,
it is assumed that the same number of draws was used for each step}

\item{prior}{prior parameters of the beta distribution of the posterior/prior counts.
The default is Jeffreys' prior.}

\item{samples}{number of samples from beta distributions used to approximate the standard error}

\item{npior}{number of prior samples that satisfy order constraints (see \code{npost})}
}
\description{
Computes the standard error of the (stepwise) encompassing Bayes factor by sampling
ratios from beta distributions, with parameters defined by the posterior/prior counts
(see Hoijtink, 2011; p. 211).
}
\examples{
npost <- 2453
nprior <- 153
M <- 10000
# Bayes factor in favor of order constraint:
npost / nprior

# standard error of estimate:
se_bf(npost, nprior, M)
}
\references{
Hoijtink, H. (2011). Informative Hypotheses: Theory and Practice for Behavioral and Social Scientists. Boca Raton, FL: Chapman & Hall/CRC.
}
