% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{select_bf}
\alias{select_bf}
\alias{select_bf.default}
\alias{select_bf.matrix}
\alias{select_bf.data.frame}
\title{Strategy Selection Using the Bayes Factor}
\usage{
select_bf(b, n, prediction, c = 0.5, prior = c(1, 1), ...)

\method{select_bf}{default}(b, n, prediction, c = 0.5, prior = c(1, 1))

\method{select_bf}{matrix}(b, n, prediction, c = 0.5, prior = c(1, 1),
  cores = 1)

select_bf.data.frame(b, n, prediction, c = 0.5, prior = c(1, 1),
  cores = 1)
}
\arguments{
\item{b}{observed frequencies of Option B.
Either a vector or a matrix/data frame (one person per row)}

\item{n}{vector with the number of choices per item type}

\item{prediction}{a vector of strategy predictions. See \code{\link{get_prediction}}}

\item{c}{upper threshold of probabilities}

\item{prior}{prior parameters of beta distribution, theta[i]~dbeta(prior[1], prior[2])}

\item{cores}{number of processing units to be used}
}
\description{
Strategy Selection Using the Bayes Factor
}
