% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nml.R
\name{select_nml}
\alias{select_nml}
\title{Strategy Selection Using NML}
\usage{
select_nml(k, n, cnml, n.fit = 5, cores = 1)
}
\arguments{
\item{k}{observed frequencies of Option B.
Either a vector or a matrix/data frame (one person per row)}

\item{n}{vector with the number of choices per item type}

\item{cnml}{NML complexity values computed with \code{\link{compute_cnml}}.
Note that the sample size \code{n} must match.}

\item{n.fit}{number of repeated fitting runs with random starting values}

\item{cores}{number of processing units to be used}
}
\description{
Compute (L)NML for observed frequencies.
}
