% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_inequalities.R
\name{sort_inequalities}
\alias{sort_inequalities}
\title{Sort Inequalities by Predictive Strength}
\usage{
sort_inequalities(A, b, M = 1e+05)
}
\arguments{
\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the binomial parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}}

\item{M}{number of uniform samples from the unit cube}
}
\description{
Uses uniform sampling to order the inequalities by their predictive strength.
The constraint that restricts the paramter space most strongly is placed at the first position.
}
\examples{
b <- c(0,0,.30,.70)
 A <- matrix(c(-1,1,0,  # p1 >= p2
               0,1,-1,  # p2 <= p3
               1,0,0,   # p1 <=.30
               0,1,0),  # p2 <= .70
               ncol = 3, byrow = 2)
sort_inequalities(A, b)
}
