% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsel_package.R
\docType{package}
\name{stratsel-package}
\alias{stratsel}
\alias{stratsel-package}
\title{stratsel:}
\description{
Outcome-based
}
\details{
Detailed explanations and examples can be found in the package vignette, accessible via \code{vignette("stratsel")}
}
\references{
Heck, D. W., Hilbig, B. E., & Moshagen, M. (2017). From information processing to decisions: Formalizing and comparing probabilistic choice models. Cognitive Psychology, 96, 26-40. https://doi.org/10.1016/j.cogpsych.2017.05.003
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danheck/stratsel}
}

}
\author{
Daniel W. Heck
}
