% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_predictions.R
\name{unique_predictions}
\alias{unique_predictions}
\title{Find Unique Predictions}
\usage{
unique_predictions(predictions, reversed = TRUE, return = "subset")
}
\arguments{
\item{predictions}{a matrix or data frame with predictions (one strategy/model per column).
Negative value = strength of preference for Option A.
Positive value = strength of preference for Option B.
0 = guessing with probability .50. See \code{\link{predict_multiattribute}}}

\item{reversed}{whether to exclude reversed patterns
(i.e., those that are identical after switching Option A and B)}

\item{return}{whether to return \code{"subset"} of input matrix of predictions or the \code{"index"} of unique lines}
}
\description{
Allows to find unique item types, which are defined as patterns of cue values
that lead to identical strategy predictions.
}
