% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_to_inequalities.R
\name{vertex_to_ineq}
\alias{vertex_to_ineq}
\alias{ineq_to_vertex}
\title{Transform Vertex/Inequality Representations of Polytopes}
\usage{
vertex_to_ineq(V)

ineq_to_vertex(A, b)
}
\arguments{
\item{V}{a matrix with one vertex of a polytope per row
(i.e., the admissible preference orders of a random utility model)}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the binomial parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}}
}
\description{
For convex polytopes: Uses \code{rPorta} (\url{https://github.com/TasCL/rPorta})
to transform the vertex representation to/from the inequality representation.
}
\details{
When defining a polytope via \code{A} and \code{b}, constraints are automatically
added to ensure that all parameters are probabilities:  0 <= x <= 1.
Note that the transformation can be very slow and might require days of computing.
}
\examples{
\dontrun{
V <- matrix(c(0, 0, 0,
              1, 0, 0,
              0, .5, 0,
              0, 0, .5,
              .5, .5, .5,
              0, .5, 0), ncol = 3, byrow = TRUE)
vertex_to_ineq(V)

A <- matrix(c(1, -1, 0,   # x1 < x2
              0, 1, -1),  # x2 < x3
            ncol = 3, byrow = TRUE)
b <- c(0, 0)
ineq_to_vertex(A, b)
}
}
