% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{wnml}
\alias{wnml}
\alias{wnml.default}
\alias{wnml.matrix}
\alias{wnml.data.frame}
\alias{post_prob.default}
\title{Get NML Model Weights}
\usage{
wnml(nml)

\method{wnml}{default}(nml)

\method{wnml}{matrix}(nml)

\method{wnml}{data.frame}(nml)

\method{post_prob}{default}(marginal, prior)
}
\arguments{
\item{nml}{vector or matrix (one row per person) with NML values}
}
\description{
Computes the NML weights for the competing models
(comparable to posterior model probabilities).
}
