% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bf.R
\name{compute_bf}
\alias{compute_bf}
\title{Encompassing Bayes Factor}
\usage{
compute_bf(k, n, A, b, prior = c(1, 1), M = 5e+05, steps, batch = 10000)
}
\arguments{
\item{k}{vector with observed frequencies for Option B}

\item{n}{vector with the number of choices per item type}

\item{A}{a matrix with one row for each linear inequality constraint and one
column for each of the binomial parameters. The parameter space is defined
as all probabilities \code{x} that fulfill the order constraints  \code{A*x <= b}.}

\item{b}{a vector of the same length as the number of rows of \code{A}}

\item{prior}{a vector with two positive numbers defining the shape parameters of the beta prior distributions for each binomial rate parameter}

\item{M}{number of posterior samples drawn from the encompassing model}

\item{steps}{integer vector that indicates at which rows the matrix \code{A} is split for a stepwise computation of the Bayes factor (see details). In this case, \code{M} can be a vector with the number of samples drawn in each step from the (partially) order-constrained models using Gibbs sampling}

\item{batch}{size of the batches into which computations are split to reduce memory load}
}
\value{
a matrix with estiamtes (first column) and standard errors (second column) for the Bayes factor and log BF of the order-constrained vs. encompassing (\code{bf_0e}) and the encompassing vs. order-constrained (\code{bf_e0}) model
}
\description{
Computes the Bayes factor for order-constrained product-binomial models
that are specified as a polytope via:  A*x <= b (see details).
}
\details{
The stepwise computation of the Bayes factor proceeds as follows: If the steps are defined as \code{steps=c(5,10)}, the BF is computed in three steps by comparing: (1) Unconstrained model vs. inequalities in \code{A[1:5,]}; (2) use posterior based on inequalities in \code{A[1:5,]} and check inequalities \code{A[6:10,]}; (3) sample from A[1:10,] and check inequalities in \code{A[11:nrow(A),]} (i.e., all inequalities).

For more control how many samples should be drawn from the prior/posterior, use \code{\link{count_polytope}}.
This is especially recommended if the same prior distribution (and thus the same integral)
is used for computing BFs for multiple data sets that differ only in \code{k} and \code{n}.
}
\examples{
k <- c(0, 3, 2, 5, 3, 7)
n <- rep(10, 6)

# linear order constraint:
# theta1 < theta2 < .... < .50
A <- matrix(c(1, -1, 0, 0, 0, 0,
              0, 1, -1, 0, 0, 0,
              0, 0, 1, -1, 0, 0,
              0, 0, 0, 1, -1, 0,
              0, 0, 0, 0, 1, -1,
              0, 0, 0, 0, 0, 1),
            ncol = 6, byrow = TRUE)
b <- c(0, 0, 0, 0, 0, .50)

# check whether vectors are in polytope:
A \%*\% c(.05, .1, .12, .16, .19, .23) <= b
A \%*\% c(.05, .3, .12, .16, .19, .53) <= b

# Bayes factor: unconstrained vs. constrained
compute_bf(k, n, A, b, c(1, 1), M = 2e5)
compute_bf(k, n, A, b, c(1, 1), M = 1e4, steps = c(2,4,5))
}
\references{
Karabatsos, G. (2005). The exchangeable multinomial model as an approach to testing deterministic axioms of choice and measurement. Journal of Mathematical Psychology, 49(1), 51-69. \url{https://doi.org/10.1016/j.jmp.2004.11.001}

Regenwetter, M., Davis-Stober, C. P., Lim, S. H., Guo, Y., Popova, A., Zwilling, C., … Messner, W. (2014). QTest: Quantitative testing of theories of binary choice. Decision, 1(1), 2-34. \url{https://doi.org/10.1037/dec0000007}
}
