% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{compute_marginal}
\alias{compute_marginal}
\title{Compute Log-Marginal Probabilities for Strategies}
\usage{
compute_marginal(k, n, prediction, prior = c(1, 1), c = 0.5)
}
\arguments{
\item{k}{observed frequencies of Option B.
Either a vector or a matrix/data frame (one person per row)}

\item{n}{vector with the number of choices per item type}

\item{prediction}{a vector of strategy predictions. See \code{\link{predict_multiattribute}}}

\item{prior}{prior parameters of beta distribution, theta[i]~dbeta(prior[1], prior[2])}

\item{c}{upper threshold of probabilities}
}
\description{
Computes the logarithm of the integral over the likelihood function weighted by the prior distribution of the parameters.
}
