% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_multinomial.R
\name{count_multinomial}
\alias{count_multinomial}
\title{Number of Product-Multinomial Prior/Posterior Samples in Polytope}
\usage{
count_multinomial(k = 0, options, A, b, V, prior = rep(1, sum(options)), M,
  steps, batch = 10000)
}
\arguments{
\item{k}{the number of choices for each alternative ordered by item type, e.g.
\code{c(a1,a2,a3, b1,b2)}.
The default \code{k=0} is equivalent to sampling from the prior.}

\item{options}{the number of choice options per item type, e.g., \code{c(3,2)} for a ternary and binary item.
The sum of \code{options} must be equal to the length of \code{k}.}

\item{A}{a matrix defining the convex polytope via A*x <= b.
The columns of A do not include the last choice option per item type and
thus the number of columns must be equal to \code{sum(options-1)} (e.g., the column order of \code{A}
for \code{k = c(a1,a2,a2, b1,b2)} is \code{c(a1,a2, b1)}).}

\item{b}{a vector of the same length as the number of rows of \code{A}.}

\item{V}{a matrix of vertices (one per row) that define the polytope of
admissible parameters as the convex hull over these points
(if provided, \code{A} and \code{b} are ignored). Note that this method
is comparatively slow since it solves a linear-programming problem for each point (Fukuda, 2004).}

\item{prior}{the prior parameters of the Dirichlet-shape parameters.
Must have the same length as \code{k}.}
}
\value{
a list with the elements
\itemize{
    \item\code{integral}: estimated probability that samples are in polytope
    \item\code{count}: number of samples in polytope
    \item\code{M}: total number of samples
}
}
\description{
Draws prior/posterior samples for product-multinomial data and counts how many samples are
inside the convex polytope defined by
(1) the inequalities A*x <= b or
(2) the convex hull over the vertices V.
}
\examples{
### frequencies:
#           (a1,a2,a3, b1,b2,b3,b4)
k <-       c(1,5,9,    5,3,7,8)
options <- c(3,        4)

### linear order constraints
# a1<a2<a3   AND   b2<b3<.50
# (note: a2<a3 <=> a2<1-a1-a2 <=> a1+2*a2<1)
# matrix A:
#             (a1,a2, b1,b2,b3)
A <- matrix(c(1, -1, 0,  0,  0,
              1,  2, 0,  0,  0,
              0,  0, 0,  1, -1,
              0,  0, 0,  0,  1),
            ncol = sum(options-1), byrow = TRUE)
b <- c(0, 1, 0, .50)

# count prior and posterior samples and get BF
prior <- count_multinomial(0, options, A, b, M = 2e5)
posterior <- count_multinomial(k, options, A, b, M = c(2e5))
posterior$integral / prior$integral  # BF for constraints
count_to_bf(posterior, prior)

bf_multinomial(k, options, A, b, M=1000)
}
\references{
Hoijtink, H. (2011). Informative Hypotheses: Theory and Practice for Behavioral and Social Scientists. Boca Raton, FL: Chapman & Hall/CRC.
}
