% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{drop_fixed}
\alias{drop_fixed}
\alias{drop_fixed.matrix}
\alias{drop_fixed.data.frame}
\alias{drop_fixed.default}
\alias{add_fixed}
\alias{add_fixed.matrix}
\alias{add_fixed.data.frame}
\alias{add_fixed.default}
\title{Drop or Add Fixed Dimensions for Multinomial Probabilities/Frequencies}
\usage{
drop_fixed(x, options = 2)

\method{drop_fixed}{matrix}(x, options = 2)

\method{drop_fixed}{data.frame}(x, options = 2)

\method{drop_fixed}{default}(x, options = 2)

add_fixed(x, options = 2, sum = 1)

\method{add_fixed}{matrix}(x, options = 2, sum = 1)

\method{add_fixed}{data.frame}(x, options = 2, sum = 1)

\method{add_fixed}{default}(x, options = 2, sum = 1)
}
\arguments{
\item{x}{a vector (typically \code{k}, \code{n}, or \code{prior}) or
a matrix (typically \code{A} or \code{V}), in which case the fixed dimension
is dropped column-wise.}

\item{options}{the number of choice options per item type,
e.g., \code{c(3,2)} for a ternary and binary item.
 The sum of \code{options} must be equal to the length of \code{k}.}

\item{sum}{a vector that determines the fixed sum in each multinomial condition.
By default, probabilities are assumed that sum to one.
If frequencies \code{n} are provided, use \code{sum=n}.}
}
\description{
Switches between two representation of polytopes for multinomial probabilities
(whether the fixed parameters are included).
}
\examples{
######## bi- and trinomial (a1,a2, b1,b2,b3)
# vectors with frequencies:
drop_fixed(c(3,7,  4,1,5), options = c(2,3))
add_fixed (c(3,    4,1),   options = c(2,3),
           sum = c(10, 10))

# matrices with probabilities:
V <- matrix(c(1,  0,  0,
              1, .5, .5,
              0,  1,  0), 3, byrow = TRUE)
V2 <- add_fixed(V, options = c(2,3))
V2
drop_fixed(V2, c(2,3))
}
