% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{get_prediction}
\alias{get_prediction}
\title{Get Predictions of Common Strategies}
\usage{
get_prediction(cueA, cueB, v, strategy)
}
\arguments{
\item{cueA}{cue values of Option A (-1/+1 = negative/positive; 0 = missing)}

\item{cueB}{cue values of Option B (see \code{cueA})}

\item{v}{cue validities: probability that cue leads to correct decision.}

\item{strategy}{strategy label, e.g., \code{"TTB"}, \code{"WADD"}, or \code{"WADDprob"}. See details.}
}
\description{
Returns a vector of predictions for well-known choice strategies (TTB, WADD)
}
