% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{get_prob_B}
\alias{get_prob_B}
\title{Get Probabilities Predicted by Strategy}
\usage{
get_prob_B(par, prediction)
}
\arguments{
\item{par}{parameter vector (i.e., error probabilities)}

\item{prediction}{a vector of strategy predictions. See details and \code{\link{predict_multiattribute}}}
}
\description{
Returns a vector with probabilities of choosing Option B for the different item types.
}
\examples{
# Predicted: B, B, [guess], A
pred <- c(1, 1, 0, -1)
# error probability = .10
get_prob_B(par = .10, pred)

# Predicted: B,B,A with errors: e1 < e2 < e3
pred2 <- c(1, 2, -3)
get_prob_B(par = c(.10, .15,.20), pred2)
}
