% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heck2017}
\alias{heck2017}
\title{Data: Multiattribute Decisions (Heck, Hilbig & Moshagen, 2017)}
\format{A data frame 4 variables:
\describe{
  \item{\code{B1}}{Frequency of choosing Option B for Item Type 1}
  \item{\code{B2}}{Frequency of choosing Option B for Item Type 2}
  \item{\code{B3}}{Frequency of choosing Option B for Item Type 3}
  \item{\code{B4}}{Frequency of choosing Option B for Item Type 4}
}}
\usage{
heck2017
}
\description{
Dataset with multiattribute decisions across 4 item types (Heck, Hilbig & Moshagen, 2017).
}
\details{
Each participant made 40 choices for each of 4 item types with four cues (with validities .9, .8, .7, and .6).
}
\examples{
data(heck2017)
head(heck2017)

\dontrun{
n <- rep(40, 4)
preds <- list(WADDD = c(-1, -1, -1, 1),
              TTB =   c(-1, -1, -1, -1),
              EQW =   c(-1, 0, -1, 1),
              GUESS = c(0, 0, 0, 0))
cnmls <- compute_cnml(preds, n, c = .5)
cnml_baseline <- compute_cnml(1:4, n, c = 1)
select_nml(heck2017[1:5,], n,
           c(list(baseline=cnml_baseline), cnmls))
}
}
\references{
Hilbig, B. E., & Moshagen, M. (2014). Generalized outcome-based strategy classification: Comparing deterministic and probabilistic choice models. Psychonomic Bulletin & Review, 21(6), 1431-1443. https://doi.org/10.3758/s13423-014-0643-0
}
\keyword{datasets}
