% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml.R
\name{maximize_ll}
\alias{maximize_ll}
\title{Maximum-likelihood Estimate}
\usage{
maximize_ll(b, n, prediction, c = 0.5, luck = c(1, 1), n.fit = 5)
}
\description{
Get ML estimate (possible weighted by luckinesss function).
}
