% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprob}
\alias{postprob}
\title{Transform Bayes Factors to Posterior Model Probabilities}
\usage{
postprob(..., prior, beta = c(0.5, 0.5), samples = 3000)
}
\arguments{
\item{...}{an arbitrary number of Bayes-factor objects with the elements \code{count} and \code{M} indicating how often posterior samples where within the admissible parameter space relative to the same encompassing model. Such lists are returned by \code{\link{count_binom}} or \code{\link{count_multinom}}.}

\item{prior}{prior model probabilities. Uniform by default.}

\item{beta}{prior shape parameters of the beta distributions used for approximating the
standard errors of the Bayes-factor estimates. The default is Jeffreys' prior.}

\item{samples}{number of samples from beta distributions used to compute
   standard errors.


The unconstrained (encompassing) model is the saturated baseline model that assumes a separate,
independent probability for each observable frequency. The Bayes factor is obtained
as the ratio of posterior/prior samples within an order-constrained subset of the
parameter space.

The standard error of the (stepwise) encompassing Bayes factor is estimated by sampling
ratios from beta distributions, with parameters defined by the posterior/prior counts
(see Hoijtink, 2011; p. 211).}
}
\description{
Computes posterior model probabilities based on objects returned by
\code{\link{bf_binom}}, \code{\link{bf_multinom}} and \code{\link{count_to_bf}}.
}
