% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_karabatsos.R, R/posterior_predictive.R
\name{ppp_binomial}
\alias{ppp_binomial}
\alias{ppp_multinomial}
\alias{ppp_binomial}
\alias{ppp_multinomial}
\title{Posterior Predictive p-Values}
\usage{
ppp_binomial(theta, k, n, by)

ppp_multinomial(theta, k, options)

ppp_binomial(theta, k, n, by)

ppp_multinomial(theta, k, options)
}
\arguments{
\item{theta}{vector with probability parameters. Must sum to one within each option.
Can be a matrix with one parameter vector per row.
See \code{\link{sampling_binomial}} and \code{\link{sampling_multinomial}}.}

\item{k}{the number of Option B choices.
The default \code{k=n=rep(0,ncol(A))} is equivalent to sampling from the prior.}

\item{n}{integer vector, specifying the total number of objects for each multinomial distribution.}

\item{by}{optional: a vector of the same length as \code{k} indicating factor levels
by which the posterior-predictive checks should be split (e.g., by item sets).}

\item{options}{number of parameters/categories of each multinomial distribution.}

\item{M}{number of subsamples from \code{theta}}

\item{by}{optional: a vector of the same length as \code{k} indicating factor levels
by which the posterior-predictive checks should be split (e.g., by item sets).}
}
\description{
Uses posterior samples to get posterior-predicted frequencies and compare
the Pearson's X^2 statistic for (1) the observed frequencies vs. (2) the posterior-predicted frequencies.

Uses posterior samples to get posterior-predicted frequencies and compare
the Pearson's X^2 statistic for (1) the observed frequencies vs. (2) the posterior-predicted frequencies.
}
\examples{
# uniform samples:  p<.10
theta <- matrix(runif(300*3, 0, .1), 300)
n <- rep(10, 3)
ppp_binomial(theta, c(1,2,0), n)  # ok
ppp_binomial(theta, c(5,4,3), n)  # misfit

# multinomial (ternary choice)
theta <- matrix(runif(300*2, 0, .05), 300)
ppp_multinomial(theta, c(1,0,9), 3)  # ok
# uniform samples:  p<.10
theta <- matrix(runif(300*3, 0, .1), 300)
n <- rep(10, 3)
ppp_binomial(theta, c(1,2,0), n)  # ok
ppp_binomial(theta, c(5,4,3), n)  # misfit

# multinomial (ternary choice)
theta <- matrix(runif(300*2, 0, .05), 300)
ppp_multinomial(theta, c(1,0,9), 3)  # ok
}
\references{
Myung, J. I., Karabatsos, G., & Iverson, G. J. (2005). A Bayesian approach to testing decision making axioms. \emph{Journal of Mathematical Psychology, 49}, 205-225. \url{https://doi.org/10.1016/j.jmp.2005.02.004}

Myung, J. I., Karabatsos, G., & Iverson, G. J. (2005). A Bayesian approach to testing decision making axioms. \emph{Journal of Mathematical Psychology, 49}, 205-225. \url{https://doi.org/10.1016/j.jmp.2005.02.004}
}
\seealso{
\code{\link{sampling_binomial}}/\code{\link{sampling_multinomial}} to get posterior samples and \code{\link{rpbinom}}/\code{\link{rpmultinom}} to get posterior-predictive samples.

\code{\link{sampling_binomial}}/\code{\link{sampling_multinomial}} to get posterior samples and \code{\link{rpbinom}}/\code{\link{rpmultinom}} to get posterior-predictive samples.
}
