% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nml.R
\name{select_nml}
\alias{select_nml}
\title{Strategy Selection Using NML}
\usage{
select_nml(k, n, strategies, n.fit = 5, cores = 1)
}
\arguments{
\item{k}{vector with observed frequencies for Option B}

\item{n}{vector with the number of choices per item type}

\item{strategies}{list with strategies including the pre-computed
NML complexity values (see \code{\link{compute_cnml}}).
Note that the sample size \code{n} must match.}

\item{n.fit}{number of repeated fitting runs with random starting values}

\item{cores}{number of processing units to be used}
}
\description{
Compute (L)NML for observed frequencies.
}
\examples{
n <- c(2, 1, 1)
# strategy: A/A/B with error probabilities {e2,e3,e1<.20}
s1 <- list(pattern = c(-3, -1, +2), c = .2,
           ordered = FALSE, prior = c(1,1))
s1n <- compute_cnml(s1, n = n)

# predict: A/A/B with ordered errors {e2<e3<e1<.50}
s2 <- list(pattern = c(-3, -1, +2), c = .5,
           ordered = TRUE, prior = c(1,1))
s2n <- compute_cnml(s2, n = n)

select_nml(c(0,1,1), n, list(s1n, s2n))
}
