% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swop.R
\docType{data}
\name{swop5}
\alias{swop5}
\title{Strict Weak Order Polytope for 5 Elements and Ternary Choices}
\format{A list with 3 elements:
\describe{
  \item{\code{A}: }{Matrix with inequality constraints that define a polytope via A*x <= b.}
  \item{\code{b}: }{vector with upper bounds for the inequalities.}
  \item{\code{start}: }{A point in the polytope.}
}}
\usage{
swop5
}
\description{
Facet-defining inequalities of the strict weak order mixture model
for the 20 paired comparisons of 5 choice elements {a,b,c,d,e}
(Regenwetter & Davis-Stober, 2012).
}
\examples{
data(swop5)
tail(swop5$A)
tail(swop5$b)
swop5$start
}
\references{
Regenwetter, M., & Davis-Stober, C. P. (2012). Behavioral variability of choices versus structural inconsistency of preferences. Psychological Review, 119(2), 408-416. \url{https://doi.org/10.1037/a0027372}
}
\keyword{datasets}
